Added provision to get maintenance part deitals.

+++ Garage.Core/Models/TrnPartUsed.cs
@@ -18,7 +18,7 @@ namespace Garage.Core.Models
         public string DocketNo { get; set; }
         public int PartID { get; set; }
         public int Qty { get; set; }
-        public float? PartCost { get; set; }
+        public double? PartCost { get; set; }
         public string PurchaseOrder { get; set; }
         public DateTime DateLogged { get; set; }
         public string LoggedBy { get; set; }
